+++ NEWS.md
@@ -4,7 +4,7 @@
 *Development Version*
 
 #### Highlights
-* 
+* Add option in SubidAttributeSummary() to summarize landuse, soil, and crop fractions for just subbasin areas (not upstream areas)
 
 ## HYPEtools 1.4.0 (2023-05-17)
 *New functions and bug fixes*